Add reverse-supervisor idea
