+++ supervisor/reverse_supervisor.py
@@ -1,0 +1,3 @@
+# the idea of this class is to be a reverse-supervisor of an evolution process
+# it has to orchestrate whole evolution with just callbacks from outside and
+# minimal feedback about the results (fitness)
